add intervention api endpoint
