+++ api/intervention/views.py
@@ -1,0 +1,14 @@
+from rest_framework import viewsets
+from plugins.snt_malaria.api.intervention.serializers import InterventionSerializer
+from plugins.snt_malaria.models import Intervention
+
+
+class InterventionViewSet(viewsets.ModelViewSet):
+    serializer_class = InterventionSerializer
+    ordering_fields = ["id", "name"]
+    http_method_names = ["get", "options"]
+
+    def get_queryset(self):
+        return Intervention.objects.filter(
+            intervention_family__account=self.request.user.iaso_profile.account
+        )
